component Picture created
