+++ react-native/training/src/app/screens/Dummy/screens/BookDetails/styles.js
@@ -1,29 +1,12 @@
 import { StyleSheet } from 'react-native';
 
-import { blue, gray, black, white } from '../../../constants/colors';
+import { blue, black, white } from '../../../../../constants/colors';
 
 const styles = StyleSheet.create({
   container: {
     flex: 1,
     alignItems: 'stretch',
     justifyContent: 'space-around'
-  },
-  imageContainer: {
-    alignItems: 'center',
-    alignSelf: 'center',
-    backgroundColor: gray,
-    borderRadius: 100,
-    height: 200,
-    justifyContent: 'center',
-    marginRight: 20,
-    width: 200
-  },
-  image: {
-    height: 200,
-    width: 200,
-    borderColor: blue,
-    borderRadius: 100,
-    borderWidth: 1
   },
   title: {
     alignSelf: 'center',
